Changed rain field in double

+++ app/src/main/java/com/tapptitude/weatherforecast/json/owm_current_weather/Rain.java
@@ -7,5 +7,5 @@ import com.google.gson.annotations.SerializedName;
  */
 public class Rain {
     @SerializedName("3h")
-    public int rain3h;
+    public double rain3h;
 }
